Standardised the way the pairwise term works

+++ mst.h
@@ -3,6 +3,6 @@
 
 #include <vector>
 
-std::vector<unsigned> sample_edge_apparence(const unsigned width, const unsigned height, const unsigned max_iter);
+std::vector<float> sample_edge_apparence(const unsigned width, const unsigned height, const unsigned max_iter);
 
 #endif // MST_H
